fixed the link between c++ and python

+++ prime_utils.cpp
@@ -3,6 +3,7 @@
 #include <iostream>
 #include <cstdio>
 #include <cstdlib>
+#include <cstdint>
 #include <vector>
 
 extern "C" bool is_prime(int n){
@@ -19,9 +20,9 @@ extern "C" bool is_prime(int n){
     return true;
 }
 
-extern "C" int count_primes(int limit){
-    int count = 0;
-    for(int i = 2; i < limit; i++){
+extern "C" int count_primes(int64_t limit){
+    int64_t count = 0;
+    for(int64_t i = 2; i < limit; i++){
         if(is_prime(i)){
             count++;
         }
@@ -29,16 +30,16 @@ extern "C" int count_primes(int limit){
     return count;
 }
 
-std::vector<int> generate_primes(int limit){
+std::vector<int> generate_primes(int64_t limit){
     std::vector<int> primes;
     if(limit < 2){
         return primes;
     }
     primes.push_back(2);
-    for (int i = 3; i <= limit; i += 2) {
+    for (int64_t i = 3; i <= limit; i += 2) {
         bool is_prime = true;
-        int max_divisor = static_cast<int>(std::sqrt(i)); 
-        for (int j = 3; j <= max_divisor; j += 2) {
+        int64_t max_divisor = static_cast<int64_t>(std::sqrt(i)); 
+        for (int64_t j = 3; j <= max_divisor; j += 2) {
             if (i % j == 0) {
                 is_prime = false;
                 break;
@@ -51,10 +52,10 @@ std::vector<int> generate_primes(int limit){
     return primes;
 }
 
-extern "C" int nth_prime(int n){
+extern "C" int nth_prime(int64_t n){
     if (n <= 0) return -1;
-    int count = 0;
-    for(int i = 2; count != n; i++){
+    int64_t count = 0;
+    for(int64_t i = 2; count != n; i++){
         if(is_prime(i)){
             count++;
         }
@@ -65,12 +66,12 @@ extern "C" int nth_prime(int n){
     return -1;
 }
 
-std::vector<int> factor(int n){
+std::vector<int> factor(int64_t n){
     std::vector<int> factors;
     if(n < 2){
         return factors;
     }
-    for(int i = 2; i <= n; i++){
+    for(int64_t i = 2; i <= n; i++){
         while(n % i == 0){
             factors.push_back(i);
             n = n / i;
@@ -79,11 +80,11 @@ std::vector<int> factor(int n){
     return factors;
 }
 
-extern "C" int previous_prime(int n){
+extern "C" int previous_prime(int64_t n){
     if(n <= 2){
         return -1;
     }
-    for(int i = n - 1; i >= 2; i--){
+    for(int64_t i = n - 1; i >= 2; i--){
         if(is_prime(i)){
             return i;
         }
@@ -91,8 +92,8 @@ extern "C" int previous_prime(int n){
     return -1;
 }
 
-extern "C" int next_prime(int n){
-    for(int i = n + 1; ; i++){
+extern "C" int next_prime(int64_t n){
+    for(int64_t i = n + 1; ; i++){
         if(is_prime(i)){
             return i;
         }
